Some clean up and refactoriing before Push
